Comments controller and Reports Controller partially implemented

+++ server/watchtower-api/Infrastructure/Repositories/ICommentRepository.cs
@@ -5,9 +5,9 @@ namespace WatchtowerApi.Infrastructure.Repositories;
 public interface ICommentRepository
 {
     Task<Comment> CreateAsync(int reportId, int userId, string commentText);
-    Task<Comment?> GetByIdAsync(int id);
-    Task<IEnumerable<Comment>> GetByReportIdAsync(long reportId);
-    Task<IEnumerable<Comment>> GetByUserIdAsync(long userId);
+    Task<Comment?> GetByIdAsync(long id);
+    Task<IEnumerable<Comment>> GetByReportIdAsync(int reportId);
+    Task<IEnumerable<Comment>> GetByUserIdAsync(int userId);
     Task<Comment> UpdateAsync(int id, string commentText);
     Task DeleteAsync(int id);
     Task<int> GetCountByReportAsync(int reportId);
